backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Network Scanning
python-nmap==0.7.1
aiohttp==3.9.1
httpx==0.26.0
beautifulsoup4==4.12.3

# Scheduling
apscheduler==3.10.4

# WebSocket Support
websockets==12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
